requirements.txt
Flask
flask-restx
Flask-SQLAlchemy
gunicorn
mysqlclient
python-dotenv
PyJWT